MY_RAG/backend/requirements.txt
fastapi
uvicorn
qdrant-client
openai
pypdf
python-dotenv
python-multipart